Rename sticker to sticker2
The new sticker is not a natural upgrade path from the original sticker, so it
does not make a lot of sense to call this 'sticker' and just bump up the
version to 2.0.0. Instead, we will give this a new name, so that both programs
can develop independently, albeit with some shared code.

The sticker2 crate inherited the version number from sticker, reset that to
0.1.0 as well.