nix: update nixpkgs
This allows us to switch to the new `libtorch-bin` derivation from
nixpkgs.

Also use join the `libtorch-bin.{dev,out}` derivations for the
`LIBTORCH` environment variable in `shell.nix`. As a result of this
change, the build script of the `tch` can correctly detect if we are
using a libtorch with CUDA support and will correctly link
`libtorch_cuda`. This leads to bigger closures, but that is not a
problem here, since `shell.nix` is only used for development.